server/requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
gunicorn
psycopg2-binary
dj-database-url